+++ src/types/TUnicodeArrayEncoding.ts
@@ -1,0 +1,1 @@
+export type TUnicodeArrayEncoding = boolean|'base64'|'rc4';